Fix craft stage
* fixed crafting stage type with multiple amounts
* made GUI datas retrievement cast-free
* fixed scoreboard not updating in some cases with /quests setStage

+++ core/src/main/java/fr/skytasul/quests/stages/StagePlayTime.java
@@ -126,7 +126,7 @@ public class StagePlayTime extends AbstractStage {
 		@Override
 		public StagePlayTime finish(LineData datas, QuestBranch branch) {
 			StagePlayTime stage = new StagePlayTime(branch);
-			stage.playTicks = (long) datas.get("ticks");
+			stage.playTicks = datas.get("ticks");
 			return stage;
 		}
